split brotli check
