+++ .circleci/config.yml
@@ -344,7 +344,6 @@ jobs:
   test-browser-interface:
     <<: *working_directory_root
     docker: *image_kernel
-
     steps:
       - <<: *attach_root
       - run:
@@ -357,6 +356,16 @@ jobs:
           working_directory: *node_workspace
           name: test
           command: make test-ci
+
+  test-browser-file-size:
+    <<: *working_directory_root
+    docker: *image_kernel
+    steps:
+      - <<: *attach_root
+      - run:
+          working_directory: *node_workspace
+          name: Check file size
+          command: npm run check-file-size
 
   build-browser-interface:
     <<: *working_directory_root
@@ -433,6 +442,7 @@ workflows:
             - build-unity-for-mac
             - build-browser-interface
             - test-browser-interface
+            - test-browser-file-size
             - test-unity-playmode
 
       - publish-desktop-artifacts:
@@ -465,4 +475,11 @@ workflows:
             - build-unity-for-webgl
             - build-browser-interface
             - test-browser-interface
+            - test-browser-file-size
             - test-unity-playmode
+
+      - test-browser-file-size:
+          <<: *all_branches
+          requires:
+            - build-unity-for-webgl
+            - build-browser-interface
